requirements.txt
Django==4.2.1
djangorestframework==3.14.0
mysqlclient==2.1.1
django-mysql==4.9.0
django-cors-headers==4.0.0
pika==1.3.2
pika-stubs==0.1.3
drf-spectacular==0.26.2
flake8==6.0.0
djangorestframework-simplejwt==5.2.2